+++ dse_do_dashboard/do_dash_app.py
@@ -9,7 +9,8 @@ from dse_do_dashboard.main_pages.home_page_edit import HomePageEdit
 from dse_do_dashboard.main_pages.prepare_data_page_edit import PrepareDataPageEdit
 from dse_do_dashboard.utils.domodelrunner import DoModelRunner, DoModelRunnerConfig
 from dse_do_utils import DataManager
-from dse_do_utils.scenariodbmanager import ScenarioDbManager
+from dse_do_utils.datamanager import Inputs, Outputs
+from dse_do_utils.scenariodbmanager import ScenarioDbManager, DatabaseType
 
 from dse_do_dashboard.dash_app import DashApp, HostEnvironment
 from dash import dcc, html, Output, Input, State
@@ -53,7 +54,7 @@ class DoDashApp(DashApp):
                  db_echo: Optional[bool] = False,
                  logo_file_name: Optional[str] = 'IBM.png',
                  navbar_brand_name: Optional[str] = 'Dashboard',
-                 cache_config: Optional[Dict]= {},
+                 cache_config: Optional[Dict] = {},
                  visualization_pages: Optional[List[VisualizationPage]]= [],
                  database_manager_class=None,
                  data_manager_class=None,
@@ -64,6 +65,9 @@ class DoDashApp(DashApp):
                  bootstrap_theme=dbc.themes.BOOTSTRAP,
                  bootstrap_figure_template:str="bootstrap",
                  enable_long_running_callbacks: bool = False,
+                 db_type: DatabaseType = DatabaseType.DB2,
+                 db_manager_kwargs: Dict = {},  # Do not set to None,
+                 dash_kwargs: Dict = {},
                  ):
         """Create a Dashboard app.
 
@@ -89,6 +93,8 @@ class DoDashApp(DashApp):
         self.db_credentials = db_credentials
         self.schema = schema
         self.db_echo = db_echo
+        self.db_type = db_type
+        self.db_manager_kwargs = db_manager_kwargs if db_manager_kwargs is not None else {}  # To ensure no None value
         self.database_manager_class = database_manager_class
         # assert issubclass(self.database_manager_class, ScenarioDbManager)
         self.dbm = self.create_database_manager_instance()
@@ -132,19 +138,24 @@ class DoDashApp(DashApp):
 
         self.job_queue: List[DoModelRunner] = []  # TODO: migrate to Store. Using global variables is dangerous
 
+
+
         super().__init__(logo_file_name=logo_file_name, navbar_brand_name=navbar_brand_name,
                          cache_config=cache_config, port=port,
                          dash_debug=dash_debug, host_env=host_env,
                          bootstrap_theme=bootstrap_theme, bootstrap_figure_template=bootstrap_figure_template,
-                         enable_long_running_callbacks=enable_long_running_callbacks,)
+                         enable_long_running_callbacks=enable_long_running_callbacks,
+                         dash_kwargs=dash_kwargs)
 
     def create_database_manager_instance(self) -> ScenarioDbManager:
         """Create an instance of a ScenarioDbManager.
         The default implementation uses the database_manager_class from the constructor.
         Optionally, override this method."""
         if self.database_manager_class is not None and self.db_credentials is not None:
-            print(f"Connecting to DB2 at {self.db_credentials['host']}, schema = {self.schema}")
-            dbm = self.database_manager_class(credentials=self.db_credentials, schema=self.schema, echo=self.db_echo)
+            print(f"Connecting to {self.db_type} at {self.db_credentials['host']}, schema = {self.schema}")
+            dbm = self.database_manager_class(credentials=self.db_credentials,
+                                              schema=self.schema, echo=self.db_echo, db_type = self.db_type,
+                                              **self.db_manager_kwargs)
         else:
             print("Error: either specifiy `database_manager_class`, `db_credentials` and `schema`, or override `create_database_manager_instance`.")
         return dbm
@@ -352,38 +363,20 @@ class DoDashApp(DashApp):
 
     def read_scenario_tables_from_db_cached(self, scenario_name: str,
                                             input_table_names: List[str] = None,
-                                            output_table_names: List[str] = None):
+                                            output_table_names: List[str] = None) -> (Inputs, Outputs):
         """For use with Flask caching. Loads data for selected input and output tables.
         Same as `read_scenario_tables_from_db`, but calls `read_scenario_table_from_db_cached`.
         Is called from dse_do_dashboard.DoDashApp to create the PlotlyManager."""
 
         if input_table_names is None:
-            # input_table_names = list(self.dbm.input_db_tables.keys())  # This is not consistent with implementation in ScenarioDbManager! Replace by empty list
             input_table_names = []
             if 'Scenario' in input_table_names: input_table_names.remove('Scenario')  # Remove the scenario table
         if output_table_names is None:  # load all tables by default
             output_table_names = self.dbm.output_db_tables.keys()
 
-        # VT 2022-09-12: Only read tables that exist in schema:
-        # TODO: test and enable this code to handle optional tables in DB, like the Warehouse, WarehouseProperties, etc
-        # input_table_names_in_schema = []
-        # for input_table_name in input_table_names:
-        #     if input_table_name in self.dbm.input_db_tables.keys():
-        #         input_table_names_in_schema.append(input_table_name)
-        #     else:
-        #         print(f"Warning: DODashApp.read_scenario_tables_from_db_cached: input table {input_table_name} not in schema. Table not read.")
-        #
-        # output_table_names_in_schema = []
-        # for output_table_name in output_table_names:
-        #     if output_table_name in self.dbm.output_db_tables.keys():
-        #         output_table_names_in_schema.append(output_table_name)
-        #     else:
-        #         print(f"Warning: DODashApp.read_scenario_tables_from_db_cached: output table {output_table_name} not in schema. Table not read.")
-
         inputs = {}
         for scenario_table_name in input_table_names:
             # print(f"read input table {scenario_table_name}")
-            # TODO: skip scenario_table_name if not in schema
             if scenario_table_name in self.dbm.input_db_tables.keys():
                 inputs[scenario_table_name] = self.read_scenario_table_from_db_cached(scenario_name, scenario_table_name)